requirements.txt
streamlit
tensorflow-cpu
opencv-python-headless
pillow
pandas
scikit-learn
matplotlib
seaborn
numpy